+++ examples/node_api_core/03_post_block.rs
@@ -5,9 +5,6 @@
 //! Submits a block as a JSON payload.
 //! Run: `cargo run --example node_api_core_post_block --release -- [NODE URL]`.
 
-use std::env;
-
-use dotenv::dotenv;
 use iota_client::{
     bee_block::{parent::Parents, Block},
     Client, Result,
@@ -18,8 +15,8 @@ async fn main() -> Result<()> {
     // Take the node URL from command line argument or use one from env as default.
     let node_url = std::env::args().nth(1).unwrap_or_else(|| {
         // This example uses dotenv, which is not safe for use in production.
-        dotenv().ok();
-        env::var("NODE_URL").unwrap()
+        dotenv::dotenv().ok();
+        std::env::var("NODE_URL").unwrap()
     });
 
     // Create a client with that node.
