+++ examples/node_api_indexer/01_get_alias_outputs.rs
@@ -1,10 +1,12 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-//! Calls `GET api/indexer/v1/outputs/basic`.
-//! Run: `cargo run --example node_api_indexer_get_basic_outputs --release -- [NODE URL] [ADDRESS]`.
+//! Calls `GET api/indexer/v1/outputs/alias`.
+//! RUn: `cargo run --example node_api_indexer_get_alias_outputs --release -- [NODE URL] [ADDRESS]`.
 
-use iota_client::{node_api::indexer::query_parameters::QueryParameter, Client, Result};
+use std::str::FromStr;
+
+use iota_client::{bee_block::output::AliasId, node_api::indexer::query_parameters::QueryParameter, Client, Result};
 
 #[tokio::main]
 async fn main() -> Result<()> {
@@ -28,23 +30,27 @@ async fn main() -> Result<()> {
         .nth(2)
         .unwrap_or_else(|| String::from("rms1qpllaj0pyveqfkwxmnngz2c488hfdtmfrj3wfkgxtk4gtyrax0jaxzt70zy"));
 
-    // Get output IDs of basic outputs that can be controlled by this address without further unlock constraints.
+    // Get output ids of outputs that can be controlled by this address.
     let output_ids = client
-        .basic_output_ids(vec![
-            QueryParameter::Address(address.to_string()),
-            QueryParameter::HasExpirationCondition(false),
-            QueryParameter::HasTimelockCondition(false),
-            QueryParameter::HasStorageReturnCondition(false),
+        .alias_output_ids(vec![
+            QueryParameter::Governor(address.to_string()),
+            QueryParameter::StateController(address.to_string()),
         ])
         .await?;
 
     // Print the address output IDs.
     println!("Address output IDs {output_ids:#?}");
 
-    // Get the outputs by their IDs.
+    // Get the outputs by their id.
     let outputs_responses = client.get_outputs(output_ids).await?;
 
     println!("Outputs: {outputs_responses:?}",);
+
+    // Get an alias output by its AliasId.
+    let alias_id = AliasId::from_str("0xd1d1e67e30effbc22671284531a5609b82969b030750468470faf03bf0afcb98")?;
+    let output_id = client.alias_output_id(alias_id).await?;
+
+    println!("Alias output: {output_id}");
 
     Ok(())
 }
