+++ examples/node_api_indexer/02_get_foundry_outputs.rs
@@ -1,10 +1,12 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-//! Calls `GET api/indexer/v1/outputs/basic`.
-//! Run: `cargo run --example node_api_indexer_get_basic_outputs --release -- [NODE URL] [ADDRESS]`.
+//! Calls `GET api/indexer/v1/outputs/foundry`.
+//! Run: `cargo run --example node_api_indexer_get_foundry_outputs --release -- [NODE URL] [ADDRESS]`.
 
-use iota_client::{node_api::indexer::query_parameters::QueryParameter, Client, Result};
+use std::str::FromStr;
+
+use iota_client::{bee_block::output::FoundryId, node_api::indexer::query_parameters::QueryParameter, Client, Result};
 
 #[tokio::main]
 async fn main() -> Result<()> {
@@ -24,27 +26,29 @@ async fn main() -> Result<()> {
         .await?;
 
     // Take the address from command line argument or use a default one.
-    let address = std::env::args()
+    let alias_address = std::env::args()
         .nth(2)
-        .unwrap_or_else(|| String::from("rms1qpllaj0pyveqfkwxmnngz2c488hfdtmfrj3wfkgxtk4gtyrax0jaxzt70zy"));
+        .unwrap_or_else(|| String::from("rms1ppdr9w5wmyg7phcd7q9exv2kvnu5rnwafftsehjpfwd6zxn83938xw83dtr"));
 
-    // Get output IDs of basic outputs that can be controlled by this address without further unlock constraints.
+    // Get output ids of foundry outputs that can be controlled by this address.
     let output_ids = client
-        .basic_output_ids(vec![
-            QueryParameter::Address(address.to_string()),
-            QueryParameter::HasExpirationCondition(false),
-            QueryParameter::HasTimelockCondition(false),
-            QueryParameter::HasStorageReturnCondition(false),
-        ])
+        .foundry_output_ids(vec![QueryParameter::AliasAddress(alias_address.to_string())])
         .await?;
 
     // Print the address output IDs.
     println!("Address output IDs {output_ids:#?}");
 
-    // Get the outputs by their IDs.
+    // Get the outputs by their id.
     let outputs_responses = client.get_outputs(output_ids).await?;
 
     println!("Outputs: {outputs_responses:?}",);
+
+    // Get an foundry output by its FoundryId.
+    let foundry_id =
+        FoundryId::from_str("0x085a32ba8ed911e0df0df00b93315664f941cddd4a570cde414b9ba11a678962730100000000")?;
+    let output_id = client.foundry_output_id(foundry_id).await?;
+
+    println!("Foundry output: {output_id}");
 
     Ok(())
 }
