+++ examples/node_api_indexer/03_get_nft_outputs.rs
@@ -1,30 +1,34 @@
 // Copyright 2022 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-//! `cargo run --example node_api_indexer_get_nft_outputs --release -- [NODE URL]`.
+//! Calls `GET api/indexer/v1/outputs/nft`.
+//! Run: `cargo run --example node_api_indexer_get_nft_outputs --release -- [NODE URL] [ADDRESS]`.
 
-use std::{env, str::FromStr};
+use std::str::FromStr;
 
-use bee_block::output::NftId;
-use dotenv::dotenv;
-use iota_client::{node_api::indexer::query_parameters::QueryParameter, Client, Result};
+use iota_client::{bee_block::output::NftId, node_api::indexer::query_parameters::QueryParameter, Client, Result};
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    // Takes the node URL from command line argument or use localhost as default.
-    let node = std::env::args().nth(1).unwrap_or_else(|| {
-        dotenv().ok();
-        env::var("NODE_URL").unwrap()
+    // Take the node URL from command line argument or use one from env as default.
+    let node_url = std::env::args().nth(1).unwrap_or_else(|| {
+        // This example uses dotenv, which is not safe for use in production.
+        dotenv::dotenv().ok();
+        std::env::var("NODE_URL").unwrap()
     });
-    // Creates a client instance with that node.
+
+    // Create a client with that node.
     let client = Client::builder()
-        // The nodes needs to have the indexer plugin enabled.
-        .with_node(&node)?
+        // The node needs to have the indexer plugin enabled.
+        .with_node(&node_url)?
         .with_node_sync_disabled()
         .finish()
         .await?;
 
-    let address = "rms1qpllaj0pyveqfkwxmnngz2c488hfdtmfrj3wfkgxtk4gtyrax0jaxzt70zy";
+    // Take the address from command line argument or use a default one.
+    let address = std::env::args()
+        .nth(2)
+        .unwrap_or_else(|| String::from("rms1qpllaj0pyveqfkwxmnngz2c488hfdtmfrj3wfkgxtk4gtyrax0jaxzt70zy"));
 
     // Get output ids of outputs that can be controlled by this address without further unlock constraints.
     let output_ids = client
@@ -36,15 +40,18 @@ async fn main() -> Result<()> {
         ])
         .await?;
 
-    println!("Address output_ids {:?}", output_ids);
+    // Print the address output IDs.
+    println!("Address output IDs {output_ids:#?}");
 
     // Get the outputs by their id.
     let outputs_responses = client.get_outputs(output_ids).await?;
+
     println!("Nft outputs: {outputs_responses:?}",);
 
     // Get an nft output by its NftId.
     let nft_id = NftId::from_str("0x649db5b14ee26d7eb91304cfeaa27cb661e1b05d366623be24d07955e0af6ce1")?;
     let output_id = client.nft_output_id(nft_id).await?;
+
     println!("Nft output: {output_id}");
 
     Ok(())
